candidates page -- sprites without text (not using :before) not working very well yet
the lines for the sprites are too large. they're showing other parts of
the background image, and I'm having to use &nbsps to get them to show
up at all, which makes weird things. I just need the <a> element with
no filler (or a single nbsp) to stay 15x15, and then I'll be fine, I
think.